pyproject.toml
[project]
name = "tender-login-actor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apify<3.0.0",
    "boto3>=1.37.18",
    "haystack-ai>=2.11.2",
    "ruff>=0.11.2",
    "selenium>=4.30.0",
    "sentence-transformers<3.0.0",
    "torch<2.3.0",
    "numpy<2.0.0",
    "pytest>=8.3.5",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
